Doc_Javelin/requirements.txt
# Core PDF manipulation
pypdf>=3.17.0

# Image processing
Pillow>=10.0.0

# Excel/Spreadsheet support
pandas>=2.0.0
openpyxl>=3.1.0

# Word document support
python-docx>=1.1.0

# Web framework
Django>=4.2.0
